backend/requirements.txt

flask
werkzeug
flask-cors
google-cloud-dialogflow
google-cloud-dialogflowcx>=1.18.0
python-dotenv
google-auth
google-auth-oauthlib
